+++ packages/implementations/react/src/composite/ColorPicker/src/components/Pointer/index.ts
@@ -1,0 +1,4 @@
+import { Pointer } from "./Pointer";
+
+export { Pointer } from "./Pointer";
+export default Pointer;