+++ packages/implementations/react/src/composite/ColorPicker/src/index.ts
@@ -1,0 +1,5 @@
+import { ColorPicker } from "./ColorPicker";
+
+export { ColorPicker } from "./ColorPicker";
+export type { ColorPickerProps, Color } from "./colorPicker.types";
+export default ColorPicker;