+++ packages/implementations/react/src/composite/Datepicker/src/components/Footer/footer.types.ts
@@ -2,6 +2,7 @@ import { ButtonHTMLAttributes } from "react";
 
 export interface DatepickerFooterProperties {
   setSelected?: () => void;
+  clear?: () => void;
 }
 
 export type DatepickerFooterProps = DatepickerFooterProperties &