+++ packages/styles/src/packages/atomic/toast/rarui-toast.css.ts
@@ -1,6 +1,10 @@
 import { keyframes, style } from "@vanilla-extract/css";
 import { recipe } from "@vanilla-extract/recipes";
-import { varsThemeBase } from "../../../themes";
+import {
+  createRainbowSprinkles,
+  defineProperties as defineRainbowProperties,
+} from "rainbow-sprinkles";
+import { mediaQueries, varsThemeBase } from "../../../themes";
 
 const toast = recipe(
   {
@@ -12,7 +16,7 @@ const toast = recipe(
       gap: varsThemeBase.spacing["2xs"],
       maxWidth: "100%",
       width: "420px",
-      zIndex: varsThemeBase.zIndex[100],
+      zIndex: varsThemeBase.zIndex[900],
     },
     variants: {
       /**
@@ -348,18 +352,37 @@ export const container = recipe(
   "container",
 );
 
-const iconContainer = style(
-  {
-    alignSelf: "self-start",
-  },
-  "icon-container",
-);
-
 export const styles = {
   toast,
   container,
   content,
-  iconContainer,
   toastContainer,
   wrapper,
 };
+
+/* -------------------------------------------------------------------------------------------------
+ * Sprinkle
+ * -----------------------------------------------------------------------------------------------*/
+
+const defineProperties = defineRainbowProperties({
+  conditions: {
+    xs: {
+      "@media": mediaQueries.xs(),
+    },
+    md: {
+      "@media": mediaQueries.md(),
+    },
+    lg: {
+      "@media": mediaQueries.lg(),
+    },
+    xl: {
+      "@media": mediaQueries.xl(),
+    },
+  },
+  defaultCondition: "md",
+  dynamicProperties: {
+    width: true,
+  },
+});
+
+export const sprinkle = createRainbowSprinkles(defineProperties);